set height to uint32

+++ transaction/chaintracker/chaintracker.go
@@ -1,5 +1,5 @@
 package chaintracker
 
 type ChainTracker interface {
-	IsValidRootForHeight(root []byte, height uint64) bool
+	IsValidRootForHeight(root []byte, height uint32) bool
 }
